+++ allennlp_models/vision/models/__init__.py
@@ -1,3 +1,4 @@
+from allennlp_models.vision.models.nlvr2 import Nlvr2Model
 from allennlp_models.vision.models.vision_text_model import VisionTextModel
 from allennlp_models.vision.models.visual_entailment import VisualEntailmentModel
 from allennlp_models.vision.models.vilbert_vqa import VqaVilbert