+++ CHANGELOG.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added some additional `__init__()` parameters to the `T5` model in `allennlp_models.generation` for customizing
   beam search and other options.
 - Added support for NLVR2 visual entailment, including a data loader, two models, and training configs.
+- Added `StanfordSentimentTreeBankDatasetReader.apply_token_indexers()` to add token_indexers rather than in `text_to_instance` 
 
 ### Fixed
 
-- Fixed registerd model name in the `pair-classification-roberta-rte` and `vgqa-vilbert` model cards.
-
+- Fixed `binary-gender-bias-mitigated-roberta-snli` model card to indicate that model requires `allennlp@v2.5.0`.
+- Fixed registered model name in the `pair-classification-roberta-rte` and `vgqa-vilbert` model cards.
 
 ## [v2.5.0](https://github.com/allenai/allennlp-models/releases/tag/v2.5.0) - 2021-06-03